Fix type cosmetics in TRUNCATE (unsigned long to unsigned int)

+++ arch/x86/64/source/boot.32.C
@@ -16,8 +16,8 @@ asm(".equ PHYS_BASE,0xFFFFFFFF00000000");
 #define TRUNCATE_A(X) (unsigned int)(((char*)X)+0x7FFFFFFF)
 #define TRUNCATE_B(X) (char*)(X+1)
 
-static inline char * truncate_b(unsigned long truncate_a_res) {
-    volatile unsigned long temp = truncate_a_res;
+static inline char * truncate_b(unsigned int truncate_a_res) {
+    volatile unsigned int temp = truncate_a_res;
     return TRUNCATE_B(temp);
 }
